rename campfire warmup to cross chain
